changes to functions and to split handlers and mutations
